feat: rename test files to match new Billboard* class names
- Rename __test__/BillBoard.spec.ts to __test__/Billboard.spec.ts
- Rename __test__/BillBoardController.spec.ts to __test__/BillboardController.spec.ts
- Rename __test__/BillBoardPlane.spec.ts to __test__/BillboardPlane.spec.ts
- Rename __test__/BillBoardOptions.spec.ts to __test__/BillboardOptions.spec.ts
- Rename __test__/BillBoardObject3D.ts to __test__/BillboardObject3D.ts
- Update import paths in renamed test files
- All 130 tests continue to pass
