+++ __test__/BillboardObject3D.ts
